+++ vs-code-extension/out/src/extension.js
@@ -253,20 +253,25 @@ function showPostEngineVersion() {
   }
 }
 
-/** Merges the post processor with any files with the '.merge.cps' extension in the same directory */
+/** Merges the post processor with any files with the '.merged.cps' extension in the same directory */
 function mergePost() {
   checkPostKernel();
   let child = require('child_process').execFile;
   let parameters = [];
   postFile = getCpsPath();
-  var mergeFile = postFile.split(".cps")[0] + ".merge.cps";
+  var mergeFile = postFile.split(".cps")[0] + ".merged.cps";
   parameters = [postFile, "--merge", mergeFile];
+
+  let includePath = vscode.workspace.getConfiguration("AutodeskPostUtility").get('includePath');
+  if (fileExists(includePath)) {
+      parameters.push("--include", includePath); // Set the include path
+  }
   try {
     var _timeout = vscode.workspace.getConfiguration("AutodeskPostUtility").get("timeoutForPostProcessing");
     _timeout *= 1000; // convert to milliseconds
-    child(postExecutable, parameters, { timeout: _timeout }, function (err, data) {
-      if (err) {
-        errorMessage("Merge failed.");
+    child(postExecutable, parameters, { timeout: _timeout }, function (err, stdout, stderr) {
+      if (stderr) {
+        errorMessage("Merge failed: " + stderr);
       } else {
         message("Merge successful. The merged post can be found in your post processors directory.");
       }
@@ -1242,14 +1247,14 @@ function setEmbeddedEslintRules() {
   let newEditorConfiguration
   switch (vscode.workspace.getConfiguration("AutodeskPostUtility").get("useEmbeddedESLintRules")) {
     case "Disabled":
-      newEditorConfiguration = Object.assign({}, currentEditorConfiguration.codeActionsOnSave, {"source.fixAll.eslint": false});
+      newEditorConfiguration = Object.assign({}, currentEditorConfiguration.codeActionsOnSave, {"source.fixAll.eslint": "never"});
       newEslintConfiguration = Object.assign({}, currentEslintConfiguration.overrideConfigFile, {});
       break;
     case "Show ESLint issues only":
-      newEditorConfiguration = Object.assign({}, currentEditorConfiguration.codeActionsOnSave, {"source.fixAll.eslint": false});
+      newEditorConfiguration = Object.assign({}, currentEditorConfiguration.codeActionsOnSave, {"source.fixAll.eslint": "never"});
       break;
     case "Show and fix ESLint issues":
-      newEditorConfiguration = Object.assign({}, currentEditorConfiguration.codeActionsOnSave, {"source.fixAll.eslint": true});
+      newEditorConfiguration = Object.assign({}, currentEditorConfiguration.codeActionsOnSave, {"source.fixAll.eslint": "explicit"});
       break;
     default:
       errorMessage("Unknown command for setting useEmbeddedESLintRules.")
@@ -1257,6 +1262,7 @@ function setEmbeddedEslintRules() {
   }
   vscode.workspace.getConfiguration("eslint").update("options", newEslintConfiguration, true);
   vscode.workspace.getConfiguration("editor").update("codeActionsOnSave", newEditorConfiguration, true);
+  vscode.workspace.getConfiguration("eslint").update("useFlatConfig", false, true);
 }
 
 /**